fix issues type
