--- src/typeScript/constants/types.ts
@@ -1,18 +1,0 @@
-import { TOAST } from "../enums";
-import { IAuth, ITodo } from "../interfaces";
-
-export type localStorageData = AuthLocalStorage | string;
-export type Toast = {
-  [key in keyof typeof TOAST]: (message: string) => string;
-};
-
-// type auth
-export type AuthLogin = Omit<IAuth, "id" | "confirmPassword">;
-export type AuthLocalStorage = Omit<IAuth, "password" | "confirmPassword"> & {
-  password?: string;
-};
-export type AuthForm = Omit<IAuth, "id">;
-
-// type todo
-export type TodoData = { title: string } | { complete: boolean };
-export type TodoItem = Omit<ITodo, "email">;
