+++ src/typeScript/types/localStorate.ts
@@ -1,0 +1,3 @@
+import { AuthLocalStorage } from "./auth";
+
+export type localStorageData = AuthLocalStorage | string;
