+++ src/typeScript/types/toast.ts
@@ -1,0 +1,5 @@
+import { TOAST } from "../enums";
+
+export type Toast = {
+  [key in keyof typeof TOAST]: (message: string) => string;
+};
