+++ src/typeScript/types/todo.ts
@@ -1,0 +1,4 @@
+import { ITodo } from "../interfaces";
+
+export type TodoData = { title: string } | { complete: boolean };
+export type TodoItem = Omit<ITodo, "email">;
